requirements.txt
streamlit==1.35.0
langchain==0.2.16
langchain-community==0.2.16
langchain-groq==0.1.9
langchain-google-genai==1.0.10
langgraph==0.2.14
chromadb==0.5.3
sentence-transformers==3.0.1
pdfplumber==0.11.1
Pillow==10.3.0
python-dotenv==1.0.1
requests==2.32.3
arxiv==2.1.3
scholarly==1.7.11
PyMuPDF==1.24.9
beautifulsoup4==4.12.3
lxml==5.2.2
google-api-python-client==2.133.0
youtube-search-python==1.6.6
groq==0.9.0
google-generativeai==0.7.2
tiktoken==0.7.0
numpy==1.26.4
pandas==2.2.2